+++ src/Kmd.Logic.Cvr.Client/CvrClient.FakeProvider.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.IO;
 using System.Threading.Tasks;
 using Kmd.Logic.Cvr.Client.Models;
 
@@ -12,7 +13,7 @@ namespace Kmd.Logic.Cvr.Client
         /// <param name="name">Name of the configuration.</param>
         /// <returns>Created configuration.</returns>
         /// <exception cref="ArgumentNullException">No parameter can be empty or null.</exception>
-        public async Task<CvrFakeProviderConfiguration> CreateFakeProviderConfiguration(string name)
+        public async Task<FakeCvrConfigurationResponseModel> CreateFakeProviderConfiguration(string name)
         {
             if (string.IsNullOrEmpty(name)) throw new ArgumentNullException(nameof(name));
 
@@ -41,7 +42,7 @@ namespace Kmd.Logic.Cvr.Client
         /// <param name="name">Name of the configuration.</param>
         /// <returns>Created configuration.</returns>
         /// <exception cref="ArgumentNullException">No parameter can be empty or null.</exception>
-        public async Task<CvrFakeProviderConfiguration> UpdateFakeProviderConfiguration(Guid configurationId, string name)
+        public async Task<FakeCvrConfigurationResponseModel> UpdateFakeProviderConfiguration(Guid configurationId, string name)
         {
             if (configurationId == Guid.Empty) throw new ArgumentNullException(nameof(configurationId));
             if (string.IsNullOrEmpty(name)) throw new ArgumentNullException(nameof(name));
@@ -51,6 +52,49 @@ namespace Kmd.Logic.Cvr.Client
                 subscriptionId: this._options.SubscriptionId,
                 configurationId: configurationId,
                 name: name).ConfigureAwait(false))
+            {
+                switch (response.Response.StatusCode)
+                {
+                    case System.Net.HttpStatusCode.OK:
+                        return response.Body;
+                    default:
+                        var responseMessage = await response.Response.Content.ReadAsStringAsync().ConfigureAwait(false);
+                        throw new CvrConfigurationException(responseMessage);
+                }
+            }
+        }
+
+        public async Task<FakeCvrConfigurationResponseModel> UpdateFakeProviderConfigurationWithData(Guid configurationId, Stream companyData, Stream productionUnitData)
+        {
+            if (configurationId == Guid.Empty) throw new ArgumentNullException(nameof(configurationId));
+
+            var client = this.CreateClient();
+            using (var response = await client.SaveDataWithHttpMessagesAsync(
+                subscriptionId: this._options.SubscriptionId,
+                configurationId: configurationId,
+                companyData: companyData,
+                productionUnitData: productionUnitData).ConfigureAwait(false))
+            {
+                switch (response.Response.StatusCode)
+                {
+                    case System.Net.HttpStatusCode.OK:
+                        return response.Body;
+                    default:
+                        var responseMessage = await response.Response.Content.ReadAsStringAsync().ConfigureAwait(false);
+                        throw new CvrConfigurationException(responseMessage);
+                }
+            }
+        }
+
+        public async Task<FakeCvrConfigurationResponseModel> UpdateFakeProviderConfigurationWithEventsData(Guid configurationId, Stream eventsData)
+        {
+            if (configurationId == Guid.Empty) throw new ArgumentNullException(nameof(configurationId));
+
+            var client = this.CreateClient();
+            using (var response = await client.SaveEventDataWithHttpMessagesAsync(
+                subscriptionId: this._options.SubscriptionId,
+                configurationId: configurationId,
+                eventsData: eventsData).ConfigureAwait(false))
             {
                 switch (response.Response.StatusCode)
                 {