+++ src/Kmd.Logic.Cvr.Client/Models/CvrProviderConfigurationModel.cs
@@ -24,12 +24,11 @@ namespace Kmd.Logic.Cvr.Client.Models
         /// Initializes a new instance of the CvrProviderConfigurationModel
         /// class.
         /// </summary>
-        public CvrProviderConfigurationModel(System.Guid? id = default(System.Guid?), System.Guid? subscriptionId = default(System.Guid?), string name = default(string), string certificateFileName = default(string), string environment = default(string), string provider = default(string))
+        public CvrProviderConfigurationModel(System.Guid? id = default(System.Guid?), System.Guid? subscriptionId = default(System.Guid?), string name = default(string), string environment = default(string), string provider = default(string))
         {
             Id = id;
             SubscriptionId = subscriptionId;
             Name = name;
-            CertificateFileName = certificateFileName;
             Environment = environment;
             Provider = provider;
             CustomInit();
@@ -54,11 +53,6 @@ namespace Kmd.Logic.Cvr.Client.Models
         /// </summary>
         [JsonProperty(PropertyName = "name")]
         public string Name { get; set; }
-
-        /// <summary>
-        /// </summary>
-        [JsonProperty(PropertyName = "certificateFileName")]
-        public string CertificateFileName { get; set; }
 
         /// <summary>
         /// </summary>