+++ CHANGES.md
@@ -1,3 +1,7 @@
+2.0.3
+* Updated CVR client for fake provoider configuration put endpoints to accept company details, production unit details
+* Updated CVR client for new endpoints added to update company events details.
+
 2.0.2
 * Extended company details and addess to accommodate changes in service contract